Move Directions files from maps to sps/exceptions and sps/model
