+++ src/main/java/com/google/sps/exceptions/DirectionsException.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.maps;
+package com.google.sps.exceptions;
 
 public class DirectionsException extends Exception {
   public DirectionsException(String message) {
