+++ src/main/java/com/google/sps/model/DirectionsClientFactory.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.maps;
+package com.google.sps.model;
 
 /**
  * Contract for creating a DirectionsClient with a given API key. Implement getDirectionsClient to
